+++ packages/expo-dev-menu/android/src/main/java/expo/modules/devmenu/compose/BindingView.kt
@@ -1,11 +1,13 @@
 package expo.modules.devmenu.compose
 
 import android.annotation.SuppressLint
+import android.app.Activity
 import android.content.Context
 import android.widget.LinearLayout
 import androidx.compose.ui.platform.ComposeView
+import expo.modules.devmenu.DevMenuManager
 import expo.modules.devmenu.compose.theme.AppTheme
-import expo.modules.devmenu.fab.ComposeMovableFloatingActionButton
+import expo.modules.devmenu.fab.MovableFloatingActionButton
 
 @SuppressLint("ViewConstructor")
 class BindingView(context: Context, lazyViewModel: Lazy<DevMenuViewModel>) : LinearLayout(context) {
@@ -18,7 +20,16 @@ class BindingView(context: Context, lazyViewModel: Lazy<DevMenuViewModel>) : Lin
         setContent {
           AppTheme {
             DevMenuScreen(viewModel.state, viewModel::onAction)
-            ComposeMovableFloatingActionButton(context, viewModel.state)
+            MovableFloatingActionButton(
+              state = viewModel.state,
+              onRefreshPress = {
+                lazyViewModel.value.onAction(DevMenuAction.Reload)
+              },
+              onOpenMenuPress = {
+                // TODO: @behenate For some reason doing onAction(DevMenuAction.Open) only works after a first refresh / opening the menu for the first time
+                DevMenuManager.openMenu(context as Activity)
+              }
+            )
           }
         }
       }